requirements.txt
pytest==6.2.3
pytest-cov==2.11.1

matplotlib~=3.4.2
networkx~=2.6.1
numpy~=1.20.2
pandas~=1.2.4
requests~=2.25.1
scipy~=1.7.0
SPARQLWrapper~=1.8.5
spacy~=3.0.6

PyDictionary~=2.0.1
PySimpleGUI~=4.47.0
pystardog=0.9.8
wordcloud~=1.8.1
word2number~=1.1